% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.value.R
\name{alpha.value}
\alias{alpha.value}
\title{Function for calculating alpha values}
\usage{
alpha.value(df)
}
\arguments{
\item{df}{must contain logMSD and Time columns}
}
\description{
This function calculates alpha values, which is the slope of logMSD v logTime. output gives both the slope and the intercept. Alpha (the slope). 2 represents perfectly ballistic migration while 1 represents random migration.
}
