% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergenceML.R
\name{convergenceML}
\alias{convergenceML}
\title{Function for determining ML convergence}
\usage{
convergenceML(dataframe)
}
\arguments{
\item{dataframe}{Time series dataframe.}
}
\description{
This function should tell us if the ML cells are converging
}
