% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacement.func.R
\name{displacement.func}
\alias{displacement.func}
\title{Function for determining displacement}
\usage{
displacement.func(df)
}
\arguments{
\item{df}{time series tracking dataframe}
}
\description{
This function calculates displacement over time. It specifically only calculates AP and ML displacement separately. For displacement in 2 dimensions, used displacement.XY
}
