% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacement.XY.R
\name{displacement.XY}
\alias{displacement.XY}
\title{Function for determining 2D displacement}
\usage{
displacement.XY(df)
}
\arguments{
\item{df}{dataframe of  X and Y values over Time}
}
\description{
This function determines 2D displacement
}
