% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftcorrect.R
\name{driftcorrect}
\alias{driftcorrect}
\title{Function for drift correction}
\usage{
driftcorrect(driftDF, rawDF, bros)
}
\arguments{
\item{driftDF}{Your dataframe for drift correction (i.e midline nuclei)}

\item{rawDF}{Raw dataframe}

\item{bros}{Number of bros in your dataframe
driftcorrect}
}
\description{
This function will drift correct the raw dataframe of X and Y coordinates collected from cell tracking. Important: both drift and raw dataframe must contain the following columns with the exact same column names: Embryo, Time, X, Y
}
