% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.somites.R
\name{label.somites}
\alias{label.somites}
\title{Function for labelling LPM cells according to the somites they are adjacent to}
\usage{
label.somites(rotatedDF, rotated.somitesDF, bros)
}
\arguments{
\item{rotatedDF}{output from rotate}

\item{bros}{number of embryos in df
label.somites()}

\item{rotate.somitesDF}{output from rotate.somites}
}
\description{
This function labels LPM cells with the somites that they are adjacent to
}
