% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLquartile.R
\name{MLquartile}
\alias{MLquartile}
\title{Function for dividing the ML positions}
\usage{
MLquartile(dataframe)
}
\arguments{
\item{dataframe}{a dataframe of drift and rotation corrected tracking values (probably output of rotate or label.somites)}
}
\description{
This function divides the limb field into 4 ML quartiles based on position, with the most medial cells being labelled 4 and the most lateral cells labelled 1. It works on each embryo seperately.
}
