% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.somites.R
\name{rotate.somites}
\alias{rotate.somites}
\title{Function for rotating the somites}
\usage{
rotate.somites(somiteDF, bros, side)
}
\arguments{
\item{somiteDF}{Dataframe containing the boundary information. must include the following column names: Embryo, X, Y, boundary. Embryo is the number of the embryo, Boundary is the somite boundary, must be in the form 0-1, 1-2, etc where the numbers represent the somites that are being bound (ie 0-1 is the boundary between somite 0 and somite 1) ALSO MUST INCLUDE minimally boundary 0-1 and 5-6!}

\item{bros}{number of embryos in the dataframe}

\item{side}{dataframe consisting of the number of the embryo, the side (L vs R) and a multiplier (-1 or 1) based on if the embryo is on the L or R side}
}
\description{
This function rotates the somites so they are aligned correctly. This is important for labelling the cells with the somite identity later
}
