% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.somite.R
\name{scatter.somite}
\alias{scatter.somite}
\title{Function for determining scatter by somite}
\usage{
scatter.somite(df, somite.start = 1, somite.final = 5)
}
\arguments{
\item{df}{time series tracking dataframe}

\item{somite.start}{starting somite that you want to determine value for, default value is 1}

\item{somite.final}{final somite you want to determine value for, default value is 5}
}
\description{
This function calculates the scatter value for each individual somite, using the scatterfun function. It calculates both AP scatter and ML scatter, and returns scatter per Embryo at each time point as well as mean scatter across all embryos of a given genotype.
}
