% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedfun.R
\name{speedfun}
\alias{speedfun}
\title{Speed function}
\usage{
speedfun(df, interval = 8)
}
\arguments{
\item{df}{cell tracking dataframe}

\item{interval}{time interval used, DEFAULT is 8 because my timepoints are all set to 8 minutes}
}
\description{
This function calculates the speed of the cells between each time point
}
