% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracklength.R
\name{tracklength}
\alias{tracklength}
\title{Function for determining the total length of a track at any point}
\usage{
tracklength(df)
}
\arguments{
\item{df}{dataframe, must contain X, Y, Time, Embryo, Track columns}
}
\description{
This function determines the total length of a track at any given time point. It does it by adding up the individual distances that each cell has moved between each time point.
}
