% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracklength.somite.R
\name{tracklength.somite}
\alias{tracklength.somite}
\title{Tracklength by somite}
\usage{
tracklength.somite(df)
}
\arguments{
\item{df}{dataframe that is the output of persitance.fun}
}
\description{
This summarizes the tracklength by somite
}
