% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistance.fun.R
\name{persistance.fun}
\alias{persistance.fun}
\title{Function for calculating persistance}
\usage{
persistance.fun(df, time)
}
\arguments{
\item{df}{dataframe}

\item{time}{final time point to calculate (frame)}
}
\description{
This function calculates persistance, which is essentially a ratio of displacement over track length. It also calculates total track length.
}
