% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterfun.R
\name{scatterfun}
\alias{scatterfun}
\title{Function for determining scatter}
\usage{
scatterfun(df)
}
\arguments{
\item{df}{time series dataframe}
}
\description{
This function calulcates scatter, which is one measure of how far apart embryos are over a time series. IT is essentially sd[t=n]/sd[t=1]
}
