% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedfun.R
\name{speedfun}
\alias{speedfun}
\title{Speed function}
\usage{
speedfun(df, time)
}
\arguments{
\item{df}{cell tracking dataframe}

\item{time}{final frame that you want to track
speedfun()}
}
\description{
This function calculates the speed of the cells between each time point
}
